# openapi.yaml
openapi: 3.1.0
info:
  title: RTK Billing API
  version: 1.0.0
  description: Provider-neutral pricing, wallet, payment, invoice, and billing access
    service.
paths:
  /v1/internal/billing/pricing-versions:
    post:
      tags:
      - Internal Billing
      summary: Create a draft pricing version and immutable rate set
      operationId: createBillingPricingVersion
      security:
      - billingServiceAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Pricing version created.
        '409':
          $ref: '#/components/responses/Error'
  /v1/internal/billing/pricing-versions/{pricingVersionId}/activate:
    post:
      tags:
      - Internal Billing
      summary: Activate one pricing version and retire the prior version
      operationId: activateBillingPricingVersion
      security:
      - billingServiceAuth: []
      parameters:
      - name: pricingVersionId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Pricing version activated.
        '404':
          $ref: '#/components/responses/Error'
  /v1/internal/billing/usage-facts:
    post:
      tags:
      - Internal Billing
      summary: Idempotently ingest a digest-anchored usage fact
      operationId: putBillingUsageFact
      security:
      - billingServiceAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Existing identical fact.
        '201':
          description: Usage fact created.
        '409':
          $ref: '#/components/responses/Error'
  /v1/internal/billing/periods/close:
    post:
      tags:
      - Internal Billing
      summary: Close, price, issue, settle, and render one billing period
      operationId: closeBillingPeriod
      security:
      - billingServiceAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Existing or newly closed invoice result.
        '422':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/billing/summary:
    get:
      tags:
      - Billing and Payments
      summary: Return tenant balance current cost runway and top-up health
      operationId: getBillingSummary
      security:
      - billingServiceAuth: []
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Tenant billing summary.
        '403':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/billing/usage:
    get:
      tags:
      - Billing and Payments
      summary: Return server-priced usage for the active billing period
      operationId: getBillingUsage
      security:
      - billingServiceAuth: []
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Priced usage lines.
        '403':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/billing/invoices:
    get:
      tags:
      - Billing and Payments
      summary: List tenant invoices with bounded filters
      operationId: listBillingInvoices
      security:
      - billingServiceAuth: []
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Invoice page.
        '403':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/billing/invoices/{invoiceId}:
    get:
      tags:
      - Billing and Payments
      summary: Get one tenant invoice and settlement links
      operationId: getBillingInvoice
      security:
      - billingServiceAuth: []
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      - name: invoiceId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Invoice detail.
        '404':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/billing/invoices/{invoiceId}/pdf:
    get:
      tags:
      - Billing and Payments
      summary: Download an immutable issued invoice PDF
      operationId: downloadBillingInvoicePdf
      security:
      - billingServiceAuth: []
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      - name: invoiceId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Immutable invoice document
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/billing/activity:
    get:
      tags:
      - Billing and Payments
      summary: List normalized customer-safe billing activity
      operationId: listBillingActivity
      security:
      - billingServiceAuth: []
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Billing activity page.
        '403':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/billing/activity/{activityId}:
    get:
      tags:
      - Billing and Payments
      summary: Get one customer-safe billing activity timeline
      operationId: getBillingActivity
      security:
      - billingServiceAuth: []
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      - name: activityId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Billing activity detail.
        '404':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/billing/profile:
    get:
      tags:
      - Billing and Payments
      summary: Get the prospective billing recipient profile
      operationId: getBillingProfile
      security:
      - billingServiceAuth: []
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Billing profile.
    put:
      tags:
      - Billing and Payments
      summary: Replace the prospective billing recipient profile
      operationId: putBillingProfile
      security:
      - billingServiceAuth: []
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      - name: If-Match
        in: header
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Updated billing profile.
        '412':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/billing/statements:
    get:
      tags:
      - Billing and Payments
      summary: Export tenant billing activity as UTF-8 CSV
      operationId: exportBillingStatement
      security:
      - billingServiceAuth: []
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: CSV statement
          content:
            text/csv:
              schema:
                type: string
  /v1/orgs/{orgId}/billing/account:
    get:
      tags:
      - Billing and Payments
      operationId: getBillingAccount
      x-rtk-feature-id: FEAT-AM-PAYMENT-001
      x-rtk-requirement-ids:
      - REQ-AM-BALANCE-LEDGER-001
      - REQ-AM-PAYMENT-AUTHZ-001
      security:
      - billingServiceAuth: []
      summary: Read the organization commercial account and automatic top-up summary.
      parameters:
      - $ref: '#/components/parameters/OrgId'
      responses:
        '200':
          description: Customer-safe commercial account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillingAccountResponse'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/billing/ledger:
    get:
      tags:
      - Billing and Payments
      operationId: listBillingLedger
      x-rtk-feature-id: FEAT-AM-PAYMENT-001
      x-rtk-requirement-ids:
      - REQ-AM-BALANCE-LEDGER-001
      - REQ-AM-PAYMENT-AUTHZ-001
      security:
      - billingServiceAuth: []
      summary: List immutable customer-safe balance ledger entries.
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Ledger page ordered newest first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillingLedgerResponse'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/payment-methods:
    get:
      tags:
      - Billing and Payments
      operationId: listPaymentMethods
      x-rtk-feature-id: FEAT-AM-PAYMENT-001
      x-rtk-requirement-ids:
      - REQ-AM-PAYMENT-METHOD-001
      - REQ-AM-PAYMENT-AUTHZ-001
      security:
      - billingServiceAuth: []
      summary: List redacted organization payment methods.
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Payment method page without provider references or card credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethodsResponse'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/payment-methods/setup:
    post:
      tags:
      - Billing and Payments
      operationId: setupPaymentMethod
      x-rtk-feature-id: FEAT-AM-PAYMENT-001
      x-rtk-requirement-ids:
      - REQ-AM-PAYMENT-METHOD-001
      - REQ-AM-PAYMENT-PROVIDER-001
      security:
      - billingServiceAuth: []
      summary: Request a provider-hosted payment method setup session.
      description: Consent, a pending method, and setup idempotency are persisted
        before provider I/O. The request never accepts PAN, CVV, expiry, or other
        card credentials. Non-production may use the isolated hosted simulator; external
        adapters remain capability-gated.
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentMethodSetupRequest'
      responses:
        '202':
          description: Hosted setup session accepted by a qualified provider.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethodSetupResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/payment-methods/{methodId}:
    delete:
      tags:
      - Billing and Payments
      operationId: revokePaymentMethod
      x-rtk-feature-id: FEAT-AM-PAYMENT-001
      x-rtk-requirement-ids:
      - REQ-AM-PAYMENT-METHOD-001
      - REQ-AM-PAYMENT-AUDIT-001
      security:
      - billingServiceAuth: []
      summary: Revoke a payment method and disable its active automatic top-up policy.
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/PaymentMethodId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentMutationReasonRequest'
      responses:
        '200':
          description: Method revoked or an identical revocation replayed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokePaymentMethodResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/auto-topup:
    get:
      tags:
      - Billing and Payments
      operationId: getAutoTopUpPolicy
      x-rtk-feature-id: FEAT-AM-PAYMENT-001
      x-rtk-requirement-ids:
      - REQ-AM-AUTO-TOPUP-001
      - REQ-AM-PAYMENT-AUTHZ-001
      security:
      - billingServiceAuth: []
      summary: Read the automatic top-up policy and Asia/Taipei safety-limit reset.
      parameters:
      - $ref: '#/components/parameters/OrgId'
      responses:
        '200':
          description: Policy response; auto_topup is null when no policy exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutoTopUpResponse'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    put:
      tags:
      - Billing and Payments
      operationId: replaceAutoTopUpPolicy
      x-rtk-feature-id: FEAT-AM-PAYMENT-001
      x-rtk-requirement-ids:
      - REQ-AM-AUTO-TOPUP-001
      - REQ-AM-PAYMENT-METHOD-001
      - REQ-AM-PAYMENT-AUDIT-001
      security:
      - billingServiceAuth: []
      summary: Replace the automatic top-up policy using optimistic concurrency and
        explicit consent.
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/IfMatchVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutAutoTopUpRequest'
      responses:
        '200':
          description: Policy replaced with a new generation and version.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutoTopUpResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '428':
          $ref: '#/components/responses/Error'
    delete:
      tags:
      - Billing and Payments
      operationId: disableAutoTopUpPolicy
      x-rtk-feature-id: FEAT-AM-PAYMENT-001
      x-rtk-requirement-ids:
      - REQ-AM-AUTO-TOPUP-001
      - REQ-AM-PAYMENT-AUDIT-001
      security:
      - billingServiceAuth: []
      summary: Disable automatic top-up and advance the policy generation.
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/IfMatchVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentMutationReasonRequest'
      responses:
        '200':
          description: Disabled policy.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutoTopUpResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '428':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/topups:
    post:
      tags:
      - Billing and Payments
      operationId: createManualTopUp
      x-rtk-feature-id: FEAT-AM-PAYMENT-001
      x-rtk-requirement-ids:
      - REQ-AM-PAYMENT-INTENT-001
      - REQ-AM-PAYMENT-AUDIT-001
      security:
      - billingServiceAuth: []
      summary: Create an idempotent manual top-up payment intent.
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualTopUpRequest'
      responses:
        '200':
          description: Existing identical intent returned for a replay.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManualTopUpResponse'
        '202':
          description: New intent persisted before provider execution.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManualTopUpResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/payment-intents:
    get:
      tags:
      - Billing and Payments
      operationId: listPaymentIntents
      x-rtk-feature-id: FEAT-AM-PAYMENT-001
      x-rtk-requirement-ids:
      - REQ-AM-PAYMENT-INTENT-001
      - REQ-AM-PAYMENT-AUTHZ-001
      security:
      - billingServiceAuth: []
      summary: List customer-safe payment intent history.
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Payment intent page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntentsResponse'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/payment-intents/{intentId}:
    get:
      tags:
      - Billing and Payments
      operationId: getPaymentIntent
      x-rtk-feature-id: FEAT-AM-PAYMENT-001
      x-rtk-requirement-ids:
      - REQ-AM-PAYMENT-INTENT-001
      - REQ-AM-PAYMENT-AUTHZ-001
      security:
      - billingServiceAuth: []
      summary: Read one customer-safe payment intent and its normalized attempts.
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/PaymentIntentId'
      responses:
        '200':
          description: Intent detail without provider transaction references or evidence
            digests.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntentDetailResponse'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /v1/internal/billing/debits:
    post:
      tags:
      - Internal Billing
      operationId: postInternalBillingDebit
      x-rtk-feature-id: FEAT-AM-PAYMENT-001
      x-rtk-requirement-ids:
      - REQ-AM-BILLING-DEBIT-001
      - REQ-AM-BALANCE-LEDGER-001
      security:
      - billingServiceAuth: []
      summary: Post one authenticated idempotent invoice or usage-adjustment debit.
      description: Uses a dedicated billing-source credential. It does not calculate
        prices or invoices and cannot post credits, manual adjustments, refunds, or
        chargebacks.
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternalBillingDebitRequest'
      responses:
        '201':
          description: Debit and balance projection committed atomically.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalBillingDebitResponse'
        '200':
          description: Semantically identical idempotent replay.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalBillingDebitResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /v1/internal/payment-simulator/setup-callback:
    post:
      tags:
      - Internal Billing
      operationId: completeSimulatedPaymentMethodSetup
      x-rtk-feature-id: FEAT-AM-PAYMENT-001
      x-rtk-requirement-ids:
      - REQ-AM-PAYMENT-METHOD-001
      - REQ-AM-PAYMENT-SIMULATOR-001
      summary: Complete a non-production simulator hosted setup using an HMAC-authenticated
        callback.
      parameters:
      - name: X-Payment-Simulator-Signature
        in: header
        required: true
        schema:
          type: string
          pattern: ^[a-f0-9]{64}$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentSimulatorSetupCallback'
      responses:
        '200':
          description: Callback accepted or replayed idempotently.
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      security: []
  /v1/payment-webhooks/{provider}:
    post:
      tags:
      - Payment Webhooks
      operationId: ingestPaymentWebhook
      x-rtk-feature-id: FEAT-AM-PAYMENT-001
      x-rtk-requirement-ids:
      - REQ-AM-PAYMENT-WEBHOOK-001
      summary: Verify, deduplicate, and enqueue a payment provider callback.
      parameters:
      - $ref: '#/components/parameters/PaymentProvider'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: string
              format: binary
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Verified callback accepted or replay deduplicated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentWebhookResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
      security: []
...
components:
  securitySchemes:
    billingServiceAuth:
      type: http
      scheme: bearer
      bearerFormat: service-token
      description: Dedicated Cloud Admin or trusted billing-worker service credential.
  schemas:
    BillingAccountResponse:
      type: object
      required:
      - account
      - auto_topup
      - payment_providers
      properties:
        account:
          $ref: '#/components/schemas/CommercialAccount'
        auto_topup:
          allOf:
          - $ref: '#/components/schemas/AutoTopUpPolicy'
          nullable: true
        payment_providers:
          type: array
          items:
            $ref: '#/components/schemas/PaymentProviderSummary'
...
  parameters:
...
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PaymentMethodId:
      name: methodId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 200
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
...
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid_claim_token:
              summary: Non-retryable Claim Token error.
              value:
                error:
                  code: invalid_claim_token
                  message: Invalid claim token
                  retryable: false
                  resolution_action: scan_or_enter_a_valid_claim_token
            claim_service_unavailable:
              summary: Retryable Claim Token service error.
              value:
                error:
                  code: service_unavailable
                  message: Claim resolve is temporarily unavailable
                  retryable: true
                  resolution_action: retry_later
            oidc_user_not_provisioned:
              summary: Unknown SSO user.
              value:
                error:
                  code: user_not_provisioned
                  message: SSO user is not provisioned in account manager
            oidc_disabled:
              summary: OIDC login disabled.
              value:
                error:
                  code: oidc_disabled
                  message: OIDC authentication is disabled
            invalid_oidc_state:
              summary: Invalid or replayed OIDC state.
              value:
                error:
                  code: invalid_oidc_state
                  message: Invalid OIDC login state
            invalid_oidc_token:
              summary: Invalid OIDC token.
              value:
                error:
                  code: invalid_oidc_token
                  message: Invalid OIDC token
            unverified_oidc_email:
              summary: Unverified OIDC email.
              value:
                error:
                  code: unverified_oidc_email
                  message: OIDC email is not verified
            oidc_provider_misconfigured:
              summary: OIDC provider misconfigured.
              value:
                error:
                  code: oidc_provider_misconfigured
                  message: OIDC provider is misconfigured